# docs/openapi/issuer-service.yaml
openapi: 3.0.3
info:
  title: CUNCTA Issuer Service
  version: 1.0.0-testnet
servers:
  - url: http://localhost:3002
paths:
...
  /v1/privacy/request:
    post:
      summary: Start DSR request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Privacy request
  /v1/privacy/confirm:
    post:
      summary: Confirm DSR request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: DSR token
  /v1/privacy/export:
    get:
      summary: Export DSR data
      security:
        - dsrToken: []
      responses:
        "200":
          description: Export payload
  /v1/privacy/restrict:
    post:
      summary: Restrict processing
      security:
        - dsrToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Restrict response
  /v1/privacy/erase:
    post:
      summary: Erase/unlink (off-chain)
      security:
        - dsrToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Erase response
components:
  securitySchemes:
    serviceJwt:
      type: http
      scheme: bearer
    accessToken:
      type: http
      scheme: bearer
    dsrToken:
      type: http
      scheme: bearer